feat: replace gos with gitea and other documentation updates
* Initial base dockerfile for devcontainer

* Add Support for VS Code Remote Containers

Add .devcontainer resources for running VSCode remote will all tools necessary to run document generation locally

Updates instructions in the CONTRIBUTING.md to explain VSCode remote support

* Update name of devcontainer

* Ignore mac .DS_Store files

* Access minikube from devcontainer

Pass env variables and mounts for kube via devcontainer.json

Add script for adjusting kube/config for accessing minikube from inside a DockerForMac container

Update the start-minikube.sh script to have minikube issue certs with docker.for.mac.localhost as a valid cert host

* Add tkn and kn cli to base devcontainer dockerfile

* Rename script for adapting kube config

Also takes in the name of a file to apply the sed transforms to

* Add .bashrc to devcontainer

Devcontainer dockerfile now pulls in a .bashrc file that can do the necessary setup for doing the tutorial in the devcontainer

Currently sets the TUTORIAL_HOME environment variable and also sets up a suitable kubeconfig if a reference can be found in the .kube directory

* Improve .bashrc

Only use the local kubeconfig if it can be adapted

Don't hardcode the location of the devcontainer kubeconfig

* Fix errant "set" in .bashrc

* Use hyperkit with minikube

Default for the latest version of minikube is to use docker.  Needs to use hyperkit instead for the tutorial to work correctly (e.g. contour routes)

* Small change to support hyperkit version of minikube

* Add platform specific options (MacOS)

hyperkit driver should only be used on Mac

don't add --apiserver-names for docker.for.mac.localhost unless minikube is running on a mac

* Fix spelling error in navigation

* Keep devcontainer files from polluting main tutorial

kubeconfig conversion script now copied into devcontainer and lives in container user home directory.

* Keep devcontainer files from polluting main tutorial

kubeconfig conversion script now copied into devcontainer and lives in container user home directory.

* Create gitea instead of gogs

Initial checkin for the creation of a gitea local github server instead of gogs

Removes the old kustomize git subdirectory

Adds gitea install yaml that has everything needed for creating the gitea operator and gitea server

* Add tekton task for initializing gitea

Converts the old gogs-init-taskrun to support equivalent setup in gitea

Adds pod/exec to list of resources the pipeline account can access so that new gitea task can be run

* Update documentation for Gitea (WIP)

First pass updating documentation for the creation and intialization of gitea repo.

NOTE: Still need to update from Trigger Template and beyond

* Update triggers documentation for Gitea (instead of gogs)

Tested with minikube, need to do a pass with OpenShift

Update triggerbinding spec for triggers 0.6.1

Hard code the context-dir to quarkus for the Greeter-trigger-template

* Update instructions to match new greeter repo

* Improve vscode dev environment

Add the kubernetes tools (and configuration) to the devcontainer

Update to the latest version of oc (and thus kubectl) in the base docker image

* Update script to support OpenShift

For now, the script install a different version of the operator (that works with OpenShift specific constructs like Routes) for OpenShift clusters

* Remove unused example files

* Fix serverless documentation for OpenShift (Namespace Prep)

There was a broken link for the OpenShift knative installation instructions

Update nexus instructions to be tabbed instead of having an optional callout for OpenShift

Fix for a couple typos

* Fix navbar order

The xrefs for triggers was in the wrong order

* Add missing CRD

The gitea CRD was missing from the gitea deployment.yaml

* Remove unused event listeners

* Split out Gitea webhook task from taskrun

Pulls the gitea webhook task from the task run so that the "WEBHOOK_HOST" can be parameterized (as this will be different in minikube compared to OpenShift)

* Add parameters to TriggerTemplate

Expose the context-dir parameter for greeter pipeline so that SpringBoot version can be (optionally) built

Expose the image-name so that this can be varied when running on OpenShift

* Update Documentation for Triggers (with Gitea) on OpenShift

For sake of documentation flow, move Triggers to the other side of the included "prepare namespace" instructions
   - also updates this to be tabbed for clearer flow with OpenShift

Add a tab with OpenShift specific instructions for creating gitea server (dealing with Routes instead of HTTPProxy)

Update instructions to allow for a parameterized EL_WEBHOOK_HOST in the gitea webhook creation
   - Gitea webhook task is now started from Tekton CLI so that parameters can be passed in

Add instructions for running springboot based greeter pipeline

Fix incorrect endpoint invocation for the (deployed) greeter service

Update reported greeter output to match documentation

Various typo fixes

* Update message to match current version of springboot based greeter

* Fix conditional function call in create-gitea.sh

Semantics of calling a function to test if target of OpenShift cluster was written incorrectly causing the script to think that everything was an openshift cluster

* Move text

Text for the persistent storage section was incorrectly left in the knative section

* Clarify Gitea introduction and explanation

Fix a couple random typos

* Add initial VSCode Remote container support (#31)

Add devcontainer support.

* Use example.org instead of example.com

example.com might have issues in certain configurations where it's already mapped in a hosts (or lmhosts) file